Select Martial Arts as the topic for the web page and add more to html base
Easier to find free images and know enough to write some information about it.
Add more to base html to overcome task since easier to just start like this and build
from there. Add classes and all that later not much format now but later on do not repeat
this probably not best practice.